共通変数に括り出していい感じにする
